[Refactor] use ign-pdal-tools + fix tests
