+++ CHANGELOG.md
@@ -1,0 +1,10 @@
+# dev
+
+- refactor: use latest version of ign-pdal-tools (v1.14.0) as a dependency
+- refactor: merge _compute_difference_with_rge_alti, _extract_rge_alti_tile_from_stream from bulk_differentiel and calcul_differentiel
+- refactor: rename calculDifferentiel to calcul_differentiel (consistency)
+- fix: temporary files usage
+
+# 1.0.0
+
+Initial version
