+++ environment.yml
@@ -3,6 +3,7 @@ channels:
   - conda-forge
 dependencies:
   - python=3.11.*
+  - gdal
   - pyproj
   - rasterio
   - pandas
@@ -10,7 +11,6 @@ dependencies:
   - tabulate
   - requests
   - pytest
-  - pysimplegui  # required for ign-gpao-utils
   # --- linting (cf https://dev.to/m1yag1/how-to-setup-your-project-with-pre-commit-black-and-flake8-183k)
   - black
   - flake8
@@ -22,5 +22,4 @@ dependencies:
     - ign-gpao-utils==0.2.2
     - ign-gpao-project-builder==0.10.0
     - ign-gpao-client==0.18.0
-    - ign-pdal-tools==1.3.0
-    - opencv-python-headless
+    - ign-pdal-tools==1.14.0
